Move Seq to utilities
